refactor: move docsApi interfaces to a separate file
